refactor: move InMemory storage to uniq package; add testify framework
