+++ datahow.go
@@ -3,48 +3,21 @@ package datahow
 import (
 	"encoding/json"
 	"net/http"
-	"sync"
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/rotiroti/datahow/uniq"
 )
-
-// A InMemory represents a concurrent-safe set of IP addresses.
-// It uses a map internally to store unique IPs.
-type InMemory struct {
-	u  map[string]struct{}
-	mu sync.Mutex
-}
-
-// New initializes a new InMemory set.
-func NewInMemory() *InMemory {
-	return &InMemory{u: make(map[string]struct{})}
-}
-
-// ExistOrAdd checks if an `ip` address exists in the set and adds it if not present.
-// Returns true if the IP was already in the set, false if it was added.
-func (i *InMemory) ExistOrAdd(ip string) bool {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-
-	if _, ok := i.u[ip]; ok {
-		return true
-	}
-
-	i.u[ip] = struct{}{}
-
-	return false
-}
 
 // LogServer represents an HTTP server that handles incoming log records.
 type LogServer struct {
 	mux    *http.ServeMux
-	store  *InMemory
+	store  *uniq.InMemory
 	metric prometheus.Counter
 }
 
 // NewLogServer creates a new LogServer with the provided InMemory storage.
-func NewLogServer(im *InMemory, c prometheus.Counter) *LogServer {
+func NewLogServer(im *uniq.InMemory, c prometheus.Counter) *LogServer {
 	srv := &LogServer{
 		store:  im,
 		metric: c,
